requirements.txt
absl-py==2.1.0
aiofiles==24.1.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
annotated-types==0.7.0
ansible==8.5.0
ansible-core==2.15.12
ansible-vault==2.1.0
anyio==4.4.0
astunparse==1.6.3
async-timeout==4.0.3
async-unzip==0.3.6
attrs==24.2.0
backoff==2.2.1
base58==2.1.1
bittensor==7.4.0
black==24.8.0
certifi==2024.7.4
cffi==1.17.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
crontab==1.0.1
cryptography==42.0.8
cytoolz==0.12.3
ddt==1.6.0
decorator==5.1.1
distlib==0.3.8
docker-pycreds==0.4.0
ecdsa==0.19.0
eth-hash==0.7.0
eth-keys==0.5.1
eth-typing==4.4.0
eth-utils==2.2.2
exceptiongroup==1.2.2
fastapi==0.110.3
filelock==3.15.4
flake8==7.1.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.6.1
fuzzywuzzy==0.18.0
gast==0.6.0
gitdb==4.0.11
GitPython==3.1.43
google-pasta==0.2.0
grpcio==1.65.5
h11==0.14.0
h5py==3.11.0
huggingface-hub==0.24.5
humanfriendly==10.0
identify==2.6.0
idna==3.7
iniconfig==2.0.0
Jinja2==3.1.4
joblib==1.4.2
keras==3.5.0
Levenshtein==0.25.1
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
ml-dtypes==0.4.0
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
msgpack-numpy-opentensor==0.5.0
multidict==6.0.5
munch==2.5.0
mypy==1.11.1
mypy-extensions==1.0.0
namex==0.0.8
nest-asyncio==1.6.0
netaddr==1.3.0
networkx==3.3
nodeenv==1.9.1
numpy==1.26.4
onnx==1.16.2
onnxruntime==1.19.0
opt-einsum==3.3.0
optree==0.12.1
packaging==24.1
password-strength==0.0.3.post2
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pre-commit==3.8.0
protobuf==4.25.4
psutil==6.0.0
py==1.11.0
py-bip39-bindings==0.1.11
py-ed25519-zebra-bindings==1.0.1
py-sr25519-bindings==0.2.0
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyflakes==3.2.0
Pygments==2.18.0
PyNaCl==1.5.0
pytest==8.3.2
pytest-asyncio==0.24.0
python-dotenv==1.0.1
python-Levenshtein==0.25.1
python-statemachine==2.1.2
PyYAML==6.0.2
rapidfuzz==3.9.6
redis==5.0.8
requests==2.32.3
resolvelib==0.8.1
retry==0.9.2
rich==13.8.0
scalecodec==1.2.11
schedule==1.2.2
scikit-learn==1.5.1
scipy==1.14.1
sentry-sdk==2.13.0
setproctitle==1.3.3
shtab==1.6.5
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
substrate-interface==1.7.10
sympy==1.13.1
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorflow==2.17.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.4.0
threadpoolctl==3.5.0
tomli==2.0.1
toolz==0.12.1
torch==2.4.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.30.0
virtualenv==20.26.4
wandb==0.17.7
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
